+++ protoc_plugin/test/all_tests.dart
@@ -33,6 +33,7 @@ import 'omit_enum_names_test.dart' as omit_enum_names;
 import 'omit_field_names_test.dart' as omit_field_names;
 import 'oneof_test.dart' as oneof;
 import 'omit_message_names_test.dart' as omit_message_names;
+import 'proto3_optional_test.dart' as proto3_optional;
 import 'protoc_options_test.dart' as protoc_options;
 import 'repeated_field_test.dart' as repeated_field;
 import 'reserved_names_test.dart' as reserved_names;
@@ -73,6 +74,7 @@ void main() {
   omit_field_names.main();
   omit_message_names.main();
   oneof.main();
+  proto3_optional.main();
   protoc_options.main();
   repeated_field.main();
   reserved_names.main();