+++ protoc_plugin/CHANGELOG.md
@@ -1,9 +1,18 @@
+## 20.0.0-nullsafety.2
+
+* Support optional proto3 fields
+
 ## 20.0.0-nullsafety.1
 
 * Generate constructors with optional named arguments for prefilling fields.
+
 ## 20.0.0-nullsafety.0
 
 * Generate null-safe code.
+
+## 19.2.1
+
+* Support optional proto3 fields.
 
 ## 19.2.0+1